Rename util to lib
As part of #105
