+++ lib/apply-migrations/index.js
