+++ lib/build-json-schema/index.js
