+++ lib/build-json-schema/util/require-id-validation.js
