+++ lib/build-json-schema/util/uniquely-identify-validation.js
