+++ lib/load-resource-models/index.js
