+++ lib/normalize-model/index.js
