+++ lib/normalize-model/util/normalize-built-in-meta.js
