+++ lib/normalize-model/util/normalize-fields.js
