+++ lib/normalize-model/util/normalize-relations.js
