+++ lib/resource-dependency-graph/index.js
