+++ lib/validate-resource-model/index.js
