+++ lib/validate-resource-model/util/resource-model-schema.json
