+++ lib/wipe-database/index.js
