+++ server/api.js
@@ -5,9 +5,9 @@ const express = require('express');
 const routeBuilder = require('express-routebuilder');
 const Resource = require('./resource');
 const serverErrors = require('./util/server-errors');
-const loadResourceModels = require('../util/load-resource-models');
-const normalizeModel = require('../util/normalize-model');
-const buildJsonSchema = require('../util/build-json-schema');
+const loadResourceModels = require('../lib/load-resource-models');
+const normalizeModel = require('../lib/normalize-model');
+const buildJsonSchema = require('../lib/build-json-schema');
 const sendJson = require('./util/send-json');
 const jsonApiHeaders = require('./util/json-api-headers');
 const createDb = require('../database');
